Fixed search & replace bug
Sometimes, when doing search-and-replace, mipscope would only replace the first instance of the search key on a given line.  Turns out the bug was not line-based, but rather cursor-based: it would push the cursor forward the length of the replaced string, right into (or over) the next thing in the line, which might have been what you were looking for in the first place.  The fix was commenting out a line that explicitly did what was just described.