nox mode now actually doesn't use X
Due to the fact that QApplication, by default, binds to an X server, even if no widgets are created. nox mode is supposed to prevent Mipscope from binding to the X server, however, as of the present version, QApplication will still bind to the X server, if present, and fail otherwise. This patch will allow Mipscope to be run without an X server by detecting if -nox is specified before creating a QApplication, and passing a flag to prevent it from binding to the X server if -nox is indeed specified.