Fix order of arguments to jalr instruction
There is a *lot* of confusion about what the correct order of arguments is.
Even Hennessy and Patterson got it wrong.  The MIPS specification (which gcc
follows) says that it's

	jalr rd, rs

where rs is the address of the instruction to execute, and rd is the
register in which to store the next instruction.  This order makes no
sense, but it is correct.