Improve input syscall handling
Aborting an input prompt (e.g. by pressing Cancel in the GUI) now throws
a syscall error that aborts the program.  This allows the user to escape
out of an infinite loop of input syscalls (the input dialog is modal,
making it impossible to access the stop button in the toolbar). [CS031 #65]

In the GUI, update the output console (i.e. flush output) before
prompting for new input.